fix: hook issue regarding actions and interactions
