+++ packages/spotlight/src/lib/interaction.ts
