+++ packages/spotlight/src/lib/mode.ts
