+++ packages/spotlight/src/modes/page/index.tsx
@@ -1,6 +1,6 @@
 import { Text } from "@mantine/core";
 
-import type { SearchMode } from "../../mode";
+import type { SearchMode } from "../../lib/mode";
 import { pagesSearchGroup } from "./pages-search-group";
 
 export const pageMode = {
