+++ packages/spotlight/src/modes/user-group/groups-search-group.tsx
@@ -3,42 +3,45 @@ import { IconEye, IconUsersGroup } from "@tabler/icons-react";
 
 import { clientApi } from "@homarr/api/client";
 
-import { createChildrenOptions } from "../../children";
-import { createGroup } from "../../group";
-import { interaction } from "../../interaction";
+import { createChildrenOptions } from "../../lib/children";
+import { createGroup } from "../../lib/group";
+import { interaction } from "../../lib/interaction";
 
 // This has to be type so it can be interpreted as Record<string, unknown>.
 // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
 type Group = { id: string; name: string };
 
 const groupChildrenOptions = createChildrenOptions<Group>({
-  useActions: [
+  useActions: () => [
     {
+      key: "details",
       component: () => (
         <Group mx="md" my="sm">
           <IconEye stroke={1.5} />
           <Text>Show group details</Text>
         </Group>
       ),
-      interaction: interaction.link(({ id }) => ({ href: `/manage/users/groups/${id}` })),
+      useInteraction: interaction.link(({ id }) => ({ href: `/manage/users/groups/${id}` })),
     },
     {
+      key: "manage-members",
       component: () => (
         <Group mx="md" my="sm">
           <IconUsersGroup stroke={1.5} />
           <Text>Manage members</Text>
         </Group>
       ),
-      interaction: interaction.link(({ id }) => ({ href: `/manage/users/groups/${id}/members` })),
+      useInteraction: interaction.link(({ id }) => ({ href: `/manage/users/groups/${id}/members` })),
     },
     {
+      key: "manage-permissions",
       component: () => (
         <Group mx="md" my="sm">
           <IconEye stroke={1.5} />
           <Text>Manage permissions</Text>
         </Group>
       ),
-      interaction: interaction.link(({ id }) => ({ href: `/manage/users/groups/${id}/permissions` })),
+      useInteraction: interaction.link(({ id }) => ({ href: `/manage/users/groups/${id}/permissions` })),
     },
   ],
   detailComponent: ({ options }) => (
@@ -53,6 +56,7 @@ const groupChildrenOptions = createChildrenOptions<Group>({
 });
 
 export const groupsSearchGroup = createGroup<Group>({
+  keyPath: "id",
   title: "Groups",
   component: ({ name }) => (
     <Group px="md" py="sm">
